SignalDecomposition.jl

SignalDecomposition.jl

Overview

Methods

Overview

SignalDecomposition.jl offers a single function:

SignalDecomposition.decompose — Function

decompose([t, ] s, method::Decomposition) → p, r

Decompose an 1D input signal or timeseries s(t) into its periodic and residual components p, r using the given method. t defaults to 1:length(s). Other common names for p, r are "seasonal" and "trend".

The decomposition is under addition, i.e. s = p .+ r, unless explicitly noted otherwise by the method. By convention, the periodic component also includes the mean value of s.

Methods

SignalDecomposition.Decomposition — Type

Abstract type of all decomposition methods.

SignalDecomposition.Fourier — Type

Fourier(frequencies) <: Decomposition

Simple decomposition method that identifies specific frequencies at the Fourier space and removes them from the signal. The residual is the inverse of the remaining Fourier signal, while the periodic part is just the removed components. If a given frequency is exactly matching the Fourier frequencies, the closest one is removed.

Important: periods/frequencies are defined with respect to the time axis length, the actual time axis t is not used in this method. So, frequency 1/12 (a period of 12) means 12 time points (whose actual time value depends on t).

This method works well when a periodic signal P is superimposed on fluctuations S, and you have a good educated guess of what frequencies compose P. This method works if the given signal has length multiple of the largest period given.

SignalDecomposition.SinusoidalFit — Type

SinusoidalFit(s, fs) <: Decomposition

Fit sinuisoidals with given frequencies fs to the signal s in decompose using the package LsqFit. Specifically, fit

\[s - \bar{s} \approx \sum_i A_i (\sin(2\pi f_i t) + \phi_i)\]

with $\bar{s}$ the mean. The fit happens on the amplitudes and phases $A_i, \phi_i$. After the decomposition you can find these in the struct's fields amps, phases. The fit is done on s versus t, so be sure that you have transformed t appropriately (e.g. if t is "days" but your frequencies are multiples of years, then you should give t/365.26).

SinusoidalFit(fs, φ0s, A0s [, ub, lb])

The quality of the fit depends dramatically on the initial guesses for the phases and amplitudes, φ0s, A0s. This second constructor gives full control over initial phases and amplitudes, as well as upper and lower bounds on the amplitudes ub, lb (also vectors of length(fs)). In the first constructor φ0 = 0, A0 = abs(-(extrema(s)...))/2, ub = Inf, lb = -Inf for all frequencies. (the bounds for the phases are always ± π)

Notice: LsqFit performs poorly for fitting sinusoidals and Fourier should be preferred over this method if the signal given is in multiples of the expected periods (and of course t is equally spaced).